Add files via upload
Made the logo into a link to the home page and adjusted the styling using the "lynx" class to keep the hover effects off of it. Tomorrow I'm going to consider resizing the array on the resumé page, as well as removing the hover effects from the nav bar to see how it will look.